refactor(pwa): rename UserManagement to PlayerManagement
